==> 02-code-scripts/r_code/01_typeI_data_prep.R <==
# this script reads the data files generated by the Scala typeI code and cleans them up.

# Load libraries
rm(list = ls(all = TRUE))
library(plyr)
library(tidyverse)

# This bit loads what we can consider as our constants for all of the analyses
source("../../02-code-scripts/r_code/common_code.R")


# there are many files to read from. The following code gets the list of files, and iteratively reads them into a data frame. This is slow!
currentwd <- getwd()
path.var <- "../../01-raw_data/typeI/"
setwd(path.var)

files <- list.files(path = path.var, pattern = "^typeI_.*\\.tsv")

read_csv_filename <- function(filename) {
  ret <- read.csv(
    filename,
    header = T,
    sep = "\t",
    skipNul = TRUE,
    stringsAsFactors = T
    # , na.strings = c("�", "∞", "-∞", "?")
  )
  ret$Source <- filename # EDIT
  ret
}

df.typeI <- ldply(files, read_csv_filename)
print("Files loaded...")

setwd(currentwd)

# done with the data read - now we can clean things up.

# First: replace the "text" symbols with the R equivelants. Order matters here as
# we want to keep our positive and negative infinity values

df.typeI <- mutate_all(df.typeI, str_replace, pattern = "-∞", replacement = "-Inf")
df.typeI <- mutate_all(df.typeI, str_replace, pattern = "∞", replacement = "Inf")
df.typeI <- mutate_all(df.typeI, str_replace, pattern = "-\\?", replacement = "NaN")
df.typeI <- mutate_all(df.typeI, str_replace, pattern = "�", replacement = "NaN")
df.typeI <- mutate_all(df.typeI, str_replace, pattern = "\\?", replacement = "NaN")

# Now that the text fields are something that R can recognise we need to convert the variables to the correct data types.
df.typeI <- df.typeI %>% mutate_at(vars(-metric, -nm, -dist, -timestamp, -Source), as.numeric)
df.typeI <- df.typeI %>% mutate_at(vars(metric, nm, dist), as.factor)
df.typeI$timestamp <- as.character(df.typeI$timestamp)

# reduce the data set to what we are interested in and remove the misc levels
df.typeI <- droplevels(subset(df.typeI, nm %in% nm_list & metric %in% ordered_metrics))

# Re-label the levels for display purposes
levels(df.typeI$dist)[levels(df.typeI$dist) == "uniform"] <- "Uniform"
levels(df.typeI$dist)[levels(df.typeI$dist) == "lognormal"] <- "Lognormal"
levels(df.typeI$dist)[levels(df.typeI$dist) == "geometric"] <- "Geometric"

# Set the order of the factors (does this hold in an RDS?)
df.typeI$nm <- factor(df.typeI$nm, levels = rev(nm_list))
df.typeI$dist <- factor(df.typeI$dist, levels = dist_list)
df.typeI$metric <- factor(df.typeI$metric, levels = ordered_metrics)

# Save it as a binary RDS file to save space and make reading it in faster
write_rds(df.typeI, file = "../../01-raw_data/typeI/df.typeI.rds", compress = "gz")


df.typeI <- readRDS(file = "../../01-raw_data/typeI/df.typeI.rds")
