==> 02-code-scripts/r_code/03_typeII_data_prep.R <==
# this script reads the data files generated by the Scala typeII code and cleans them up.

# Load libraries
library(plyr)
library(tidyverse)
library(viridis)
source("../../02-code-scripts/r_code/common_code.R")

currentwd <- getwd()
path.var <- "../../01-raw_data/typeII"
setwd(path.var)

files <- list.files(path = path.var, pattern = "^typeII_.*\\.tsv")

read_csv_filename <- function(filename) {
  ret <- read.csv(
    filename,
    header = T,
    sep = "\t",
    skipNul = TRUE,
    # na.strings = c("�", "∞", "-∞", "?"),
    stringsAsFactors = T
  )
  ret$Source <- filename # EDIT
  ret
}

df.typeII <- ldply(files, read_csv_filename)

setwd(currentwd)

# done with the data read - now we can clean things up.

# First: replace the "text" symbols with the R equivelants. Order matter here as
# we want to keep our positive and negative infinity values

df.typeII <- mutate_all(df.typeII, str_replace, pattern = "-∞", replacement = "-Inf")
df.typeII <- mutate_all(df.typeII, str_replace, pattern = "∞", replacement = "Inf")
df.typeII <- mutate_all(df.typeII, str_replace, pattern = "-\\?", replacement = "NaN")
df.typeII <- mutate_all(df.typeII, str_replace, pattern = "�", replacement = "NaN")
df.typeII <- mutate_all(df.typeII, str_replace, pattern = "\\?", replacement = "NaN")

# Now that the text fields are something that R can recognise we need to convert the variables to the correct data types.
df.typeII <- df.typeII %>% mutate_at(vars(-metric, -nm, -dist, -dir, -timestamp, -Source), as.numeric)
df.typeII <- df.typeII %>% mutate_at(vars(metric, nm, dist, dir), as.factor)

# reduce the data set to what we are interested in and remove the misc levels
df.typeII <- droplevels(subset(df.typeII, nm %in% nm_list & metric %in% ordered_metrics))

levels(df.typeII$dist)[levels(df.typeII$dist) == "uniform"] <- "Uniform"
levels(df.typeII$dist)[levels(df.typeII$dist) == "lognormal"] <- "Lognormal"
levels(df.typeII$dist)[levels(df.typeII$dist) == "geometric"] <- "Geometric"

# Set the order of the factors (does this hold in an RDS?)
df.typeII$nm <- factor(df.typeII$nm, levels = rev(nm_list))
df.typeII$dist <- factor(df.typeII$dist, levels = dist_list)
df.typeII$metric <- factor(df.typeII$metric, levels = ordered_metrics)

# Save it as a binary RDS file to save space and make reading it in faster
write_rds(df.typeII, file = "../../01-raw_data/typeII/df.typeII.rds", compress = "gz")


df.typeII <- readRDS(file = "../../01-raw_data/typeII/df.typeII.rds")
