==> 02-code-scripts/r_code/05_noise_data_prep.R <==
library(plyr)
library(tidyverse)
library(viridis)
source("../../02-code-scripts/r_code/common_code.R")

# Read in the parsed data generated by noiseParser.scala class
df.noise <- read.csv("../../01-raw_data/noise_parsed/noise_parsed.tsv", sep = "\t")

# First: replace the "text" symbols with the R equivalents. Order matter here as
# we want to keep our positive and negative infinity values

df.noise <- mutate_all(df.noise, str_replace, pattern = "-∞", replacement = "-Inf")
df.noise <- mutate_all(df.noise, str_replace, pattern = "∞", replacement = "Inf")
df.noise <- mutate_all(df.noise, str_replace, pattern = "-\\?|�|\\?", replacement = "NaN")
# df.noise <- mutate_all(df.noise, str_replace, pattern = "�", replacement = "NaN")
# df.noise <- mutate_all(df.noise, str_replace, pattern = "\\?", replacement = "NaN")

# Now that the text fields are something that R can recognise we need to convert the variables to the correct data types.
df.noise <- df.noise %>% mutate_at(vars(-mtxid, -metric, -nm, -dist, -dir, -timestamp), as.numeric)
df.noise <- df.noise %>% mutate_at(vars(mtxid, metric, nm, dist, dir), as.factor)

# reduce the data set to what we are interested in and remove the misc levels
df.noise <- droplevels(subset(df.noise, nm %in% nm_list & metric %in% ordered_metrics))

# Now calculate the percentage noise and add it to the data frame
df.noise$percent.step <- (df.noise$step - 5) / ((df.noise$species * df.noise$plots) / 2)

# Make things "pretty" for plotting
levels(df.noise$dist)[levels(df.noise$dist) == "uniform"] <- "Uniform"
levels(df.noise$dist)[levels(df.noise$dist) == "lognormal"] <- "Lognormal"
levels(df.noise$dist)[levels(df.noise$dist) == "geometric"] <- "Geometric"

levels(df.noise$dir)[levels(df.noise$dir) == "negative"] <- "Maximal Negative"
levels(df.noise$dir)[levels(df.noise$dir) == "positive"] <- "Maximal Positive"

# Set the order of the factors (does this hold in an RDS?)
df.noise$nm <- factor(df.noise$nm, levels = rev(nm_list))
df.noise$dist <- factor(df.noise$dist, levels = dist_list)
df.noise$metric <- factor(df.noise$metric, levels = ordered_metrics)

# Save it as a binary RDS file to save space and make reading it in faster
write_rds(df.noise, file = "../../01-raw_data/noise/df.noise.rds", compress = "gz")


df.noise <- readRDS(file = "../../01-raw_data/noise/df.noise.rds")
